Cursory implementation of Lexer...
for some reason my changes to tokenizer didn't post in the last commit; sorry about that.
Lexer currently outputs lexemes that it recognizes... since there are a lot of them we aren't implementing, the files that we wrote for previous labs don't work.  Tonight I'm going to go through and remove the instructions we don't implement.